+++ src/main/java/com/weather/weatherfcst/WeatherFcstApplication.java
@@ -1,4 +1,4 @@
-package com.midfcst.weatherfcst;
+package com.weather.weatherfcst;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
