+++ src/main/resources/META-INF/aop.xml
@@ -1,0 +1,8 @@
+<aspectj>
+  <weaver>
+    <include within="com.weather.weatherfcst..*"/>
+  </weaver>
+  <aspects>
+    <aspect name="com.weather.weatherfcst.aop.LoggingAspect"/>
+  </aspects>
+</aspectj>
